+++ Changes.md
@@ -3,6 +3,9 @@
 - Relaxed some dependencies for the benefit of packaging precious for
   Debian. Implemented by Jonas Smedegaard.
 
+- Added support for `.precious.toml` as a config name. Based on #21, by Olaf
+  Alders. Fixes #13.
+
 ## 0.1.2 - 2021-10-14
 
 - The order of commands in the config file is now preserved, and commands are